requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dict-toolbox==5.0.0
executing==2.2.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
frozenlist==1.6.0
idna==3.7
inference-sdk==0.47.0
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pillow_heif==0.22.0
platformdirs==4.3.6
pop==27.1.0
pop-config==12.0.4
pop-loop==1.1.0
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.2.1
rend==7.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.61
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
supervision==0.25.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.0
